+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Update app layout known static assets [#1]
+
 ## [0.2.2] - 2025-07-03
 
 - Fix manifest references for production environment [#4]